refactor(cli): extract session.py UI prints to _session_ui
B4 of the Round 2 plan. Moves the three Rich-styled print helpers
(_print_tab_bar, _print_namespace_hint, _print_session_help) and the
_TAB_ORDER constant out of amx/cli_support/session.py into a focused
amx/cli_support/_session_ui.py.

The trio renders the namespace tab bar, the per-namespace hint, and
the very long per-namespace /help text. _print_session_help alone is
366 LOC. _canonical_namespace (a 2-line helper they share) is
duplicated in the new module to avoid a circular import.

session.py re-exports all four names. session.py: 1580 -> 1168
LOC (-412, -26%).
